Higher priority for gem assets in project folders
Prior to this change, if a Gem was placed inside a Project folder, the asset paths in the cache would be relative to the Project folder, which usually breaks any assets that were using relative paths.  This change makes those Gem assets relative to the Gem folder like they would be if the Gem was outside the Project folder.  The caveat is these Gem assets in Project folders will not be overridden by assets in the Project folder, though that is a much less common use case.

Signed-off-by: Alex Peterson <26804013+AMZN-alexpete@users.noreply.github.com>